pyproject.toml
[project]
name = "funmap"
version = "0.2.0"
description = "generate gene co-function networks using omics data"
authors = [{ name = "Zhiao Shi", email = "zhiao.shi@gmail.com" }]
dependencies = [
    "PyPDF2==3.0.1",
    "click==8.1.7",
    "h5py==3.10.0",
    "imbalanced-learn==0.11.0",
    "joblib==1.3.2",
    "matplotlib==3.7.3",
    "matplotlib_venn==0.11.9",
    "networkx==3.1",
    "numpy==1.24.4",
    "pandas==2.0.3",
    "powerlaw==1.5",
    "pyarrow==13.0.0",
    "pyyaml==6.0.1",
    "scikit-learn==1.3.2",
    "scipy==1.10.1",
    "seaborn==0.13.0",
    "tables==3.8.0",
    "tqdm==4.66.1",
    "xgboost==2.0.0",
]
readme = "README.md"
license = { text = "MIT license" }
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]

[build-system]
requires = ["maturin>=1,<2"]
build-backend = "maturin"

[tool.maturin]
features = ["pyo3/extension-module"]
python-source = "python"
module-name = "funmap._lib"

[tool.rye]
managed = true
dev-dependencies = ["maturin>=1.7.4"]
